Move RepositoryStructureInspector into GitRepository assembly
refs EP-24702
