+++ Bluewire.Tools.Runner/FindBuild/ToolRunner.cs
@@ -10,6 +10,7 @@ using Bluewire.Conventions;
 using log4net.Core;
 using Bluewire.Tools.Builds.FindBuild;
 using Bluewire.Tools.Builds.Shared;
+using Bluewire.Tools.GitRepository;
 
 namespace Bluewire.Tools.Runner.FindBuild
 {
